requirements.txt
requests
pandas
beautifulsoup4
rapidfuzz
transformers
wikipedia
spacy
google-generativeai
